combined @users and @lockoutReasons modules into a single @identity module
