+++ src/identity/index.ts
@@ -1,0 +1,4 @@
+export * from '@identity/keys.ts';
+export * from '@identity/setup.tsx';
+export { default as identitySetup } from '@identity/setup.tsx';
+export type * from '@identity/types.ts';
