--- src/lockout-reasons/index.ts
@@ -1,2 +1,0 @@
-﻿export * from './keys.ts';
-export { default as lockoutReasonsSetup } from './setup.tsx';
