--- src/users/index.ts
@@ -1,2 +1,0 @@
-﻿export * from '@users/keys.ts';
-export { default as usersSetup } from '@users/setup.tsx';
